# gym-api-specification.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Gym CRM System",
    "description": "API Definitions of the Gym crm Microservice",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:8081",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/account/change-password": {
      "put": {
        "tags": [
          "Account endpoint"
        ],
        "operationId": "changePassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordChangeDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
    "/account/login": {
      "post": {
        "tags": [
          "Account endpoint"
        ],
        "summary": "Logs the user in",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
    "/account/{username}/status": {
      "patch": {
        "tags": [
          "Account endpoint"
        ],
        "operationId": "setIsActiveStatus",
        "parameters": [
          {
            "name": "username",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "active",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "PasswordChangeDto": {
        "required": [
          "newPassword",
          "oldPassword",
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "oldPassword": {
            "maxLength": 10,
            "minLength": 10,
            "type": "string"
          },
          "newPassword": {
            "maxLength": 10,
            "minLength": 10,
            "type": "string"
          }
        }
      },
...
      "LoginDto": {
        "required": [
          "password",
          "username"
        ],
        "type": "object",
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
    }
  }
}